feat: make meta/feedback agents agentic loops (like siawdh)
Replace one-shot LLM calls with full agentic loops (write_file, read_file,
bash tools) for both meta and feedback agents — mirroring siawdh's pattern.

Changes:
- Add _run_agentic_loop() shared helper (OpenAI-compatible tool-calling loop)
- Meta agent: reads reference_target_agent.py + task.md from disk, writes
  the full target_agent.py via write_file tool (6 turns max)
- Feedback agent: reads context.md, results.json, prev target_agent.py,
  sol.py, agent_execution.json from disk, writes improvement.md +
  target_agent.py via write_file tool (8 turns max)
- reference_target_agent.py: saves agent_execution.json at end of run so
  the feedback agent can inspect the full tool call trace
- Remove _inject_system_prompt/_extract_system_prompt/_clean_agent_output
  helpers (no longer needed — agent writes the full script directly)
- evaluate.py dtype fix already in this branch (float16 → float32)

+++ tasks/gpu_kernel_task/evaluate.py
@@ -264,7 +264,7 @@ def run_tensara_checker(
     """Check correctness against the reference implementation before benchmarking."""
     print(f"Checking correctness on Tensara for problem '{slug}' (language: {language})...")
     try:
-        stream = client.run_checker(slug, code, dtype="float16", language=language, gpu_type=gpu_type)
+        stream = client.run_checker(slug, code, dtype="float32", language=language, gpu_type=gpu_type)
     except Exception as e:
         return {"status": "CLIENT_ERROR", "errorMessage": str(e), "test_results": []}
     try:
@@ -279,7 +279,7 @@ def run_tensara_benchmark(
     """Run benchmark on Tensara and parse streaming SSE results."""
     print(f"Running benchmark on Tensara for problem '{slug}' (language: {language})...")
     try:
-        stream = client.run_benchmark(slug, code, dtype="float16", language=language, gpu_type=gpu_type)
+        stream = client.run_benchmark(slug, code, dtype="float32", language=language, gpu_type=gpu_type)
     except Exception as e:
         return {"status": "CLIENT_ERROR", "errorMessage": str(e), "test_results": []}
     try:
